refactor(p4-4/4): tenant/role WHERE-builders → campaignScope.js (P4-4)
